FoodBankHub/requirements.txt
# Core framework and config
Django==5.2.7
python-decouple==3.8
gunicorn==23.0.0
whitenoise==6.8.2
psycopg[binary]==3.2.3

# Feature dependencies
Pillow==12.0.0
openpyxl==3.1.2
reportlab==4.2.2
stripe==7.12.0
weasyprint==62.3
XlsxWriter==3.1.9

# Django runtime dependencies (pinned)
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2025.2

